feat: common hook moved from pages dir
