+++ frontend/hooks/languageHook.tsx
@@ -1,11 +1,11 @@
 // svgs
-import { ReactComponent as FlagSpain } from "../../assets/svg/files/españa.svg";
-import { ReactComponent as FlagUSA } from "../../assets/svg/files/usa.svg";
-import { ReactComponent as FlagItaly } from "../../assets/svg/files/italia.svg";
-import { ReactComponent as BrazilFlag } from "../../assets/svg/files/brazil.svg";
+import { ReactComponent as FlagSpain } from "@/assets/svg/files/españa.svg";
+import { ReactComponent as FlagUSA } from "@/assets/svg/files/usa.svg";
+import { ReactComponent as FlagItaly } from "@/assets/svg/files/italia.svg";
+import { ReactComponent as BrazilFlag } from "@/assets/svg/files/brazil.svg";
 //
 import { useEffect, useState } from "react";
-import i18n from "../../i18n";
+import i18n from "@/i18n";
 
 export const LanguageHook = () => {
   const [languageMenu, setLanguageMenu] = useState<any>({
