+++ frontend/pages/home/components/AddAsset.tsx
@@ -7,7 +7,7 @@ import { GeneralHook } from "../hooks/generalHook";
 import { AccountDefaultEnum, AddingAssetsEnum, TokenNetworkEnum } from "@/const";
 import { TokenHook } from "../hooks/tokenHook";
 import { Asset } from "@redux/models/AccountModels";
-import { AccountHook } from "@pages/hooks/accountHook";
+import { AccountHook } from "@/hooks/accountHook";
 import { Token } from "@redux/models/TokenModels";
 import { AssetHook } from "../hooks/assetHook";
 import { useAppDispatch } from "@redux/Store";
