+++ frontend/pages/home/components/AssetsList.tsx
@@ -9,7 +9,7 @@ import AddAsset from "./AddAsset";
 import { DrawerHook } from "../hooks/drawerHook";
 import { useTranslation } from "react-i18next";
 import Menu from "@pages/components/Menu";
-import { WorkerHook } from "@pages/hooks/workerHook";
+import { WorkerHook } from "@/hooks/workerHook";
 import { AssetHook } from "../hooks/assetHook";
 import { UseAsset } from "../hooks/useAsset";
