+++ frontend/pages/home/components/DeleteAssetModal.tsx
@@ -3,7 +3,7 @@ import { ReactComponent as WarningIcon } from "@assets/svg/files/warning.svg";
 import { ReactComponent as CloseIcon } from "@assets/svg/files/close.svg";
 //
 import Modal from "@components/Modal";
-import { AccountHook } from "@pages/hooks/accountHook";
+import { AccountHook } from "@/hooks/accountHook";
 import { useTranslation } from "react-i18next";
 import { Token } from "@redux/models/TokenModels";
 import { CustomButton } from "@components/Button";
